prometheus: Guard against successfull empty metadata requests
When a Loki instance is configured as Prometheus datasource the call to
fetch the metrics name by calling `fetchLabelValues('__name__')` returns
a successfull response but without a `data` attribute in the body
breaking the query editor when `processHistogramMetrics` is called.

If no `data` property is found in the response from the server
`defaultValue` is returned instead.